# services/gateway/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Food Delivery Backend API - Restaurants, orders, delivery",
        "title": "Food Delivery API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/websocket/connections": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "websocket"
                ],
                "summary": "List WebSocket connections",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object"
                        }
                    }
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}